Skip screenshot of first position and quit on getting stuck
The script now skips the first position, assuming that you have already
taken a shot for it on a previous run. It also now quits when it gets
stuck instead of retrying. PhantomJS is too flaky to be long running, so
by combining these features with a batch file which runs roadTripper a
few dozen times I've been able to get it taking screenshots for longer
periods of time.

To force the script to actually take the first screenshot instead of
skipping it -- e.g. after a reposition -- call the first run of
roadtripper with the argument *noskipfirst*.